Change Game to Survival
The new game works like this:
Every time the mouse is clicked, a new random shape is generated.

If the shape collides with another shape, the one with a larger first side length "takes over" the smaller shape and changes its color.

If a shape does not take anybody over for 1000 game ticks, then the shape dies.

press r to restart